``refactor``: Refactor view related function to its own Class
