+++ core/Router.php
@@ -30,7 +30,7 @@ class Router {
             throw new NotFoundException();
         }
         if (is_string($instance)) {
-            return $this->renderView($instance);
+            return Application::view()->renderView($instance);
         }
         if (is_array($instance)) {
             $instance[0] = new $instance[0]();
@@ -59,26 +59,5 @@ class Router {
 
             $middleware->execute($this->request);
         }
-    }
-
-    public function renderView(string $name, array $props = []): string {
-        $content = $this->getContent($name, $props);
-        $layout = $this->getLayout();
-        return str_replace('{{ CONTENT }}', $content, $layout);
-    }
-
-    private function getLayout(): string {
-        ob_start();
-        @require_once Application::$ROOT_DIR . '/view/app.sapi.php';
-        return ob_get_clean();
-    }
-
-    private function getContent(string $name, array $props): string {
-        foreach ($props as $key => $value) {
-            $$key = $value;
-        }
-        ob_start();
-        @require_once Application::$ROOT_DIR . "/view/{$name}.sapi.php";
-        return ob_get_clean();
     }
 }
